fix: authenticator lumber unit tests
